+++ app/Http/Requests/StorePermissionsRequest.php
@@ -1,9 +1,12 @@
-<?php
-namespace App\Http\Requests\Admin;
+<?php namespace App\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class StoreRolesRequest extends FormRequest
+/**
+ * Class StorePermissionsRequest
+ * @package App\Http\Requests
+ */
+class StorePermissionsRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
