+++ app/Http/Requests/StoreProjectStatusesRequest.php
@@ -1,9 +1,12 @@
-<?php
-namespace App\Http\Requests\Admin;
+<?php namespace App\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class StoreUsersRequest extends FormRequest
+/**
+ * Class StoreProjectStatusesRequest
+ * @package App\Modules\Admin\Requests
+ */
+class StoreProjectStatusesRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -23,10 +26,7 @@ class StoreUsersRequest extends FormRequest
     public function rules()
     {
         return [
-            'name' => 'required',
-            'email' => 'required|email|unique:users,email',
-            'password' => 'required',
-            'roles' => 'required'
+            'title' => 'required',
         ];
     }
 }
