+++ app/Http/Requests/StoreProjectsRequest.php
@@ -1,0 +1,33 @@
+<?php namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+/**
+ * Class StoreProjectsRequest
+ * @package App\Modules\Admin\Requests
+ */
+class StoreProjectsRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            'client_id' => 'required',
+            'project_status_id' => 'required',
+        ];
+    }
+}
