+++ app/Http/Requests/UpdateProjectStatusesRequest.php
@@ -1,0 +1,32 @@
+<?php namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+/**
+ * Class UpdateProjectStatusesRequest
+ * @package App\Http\Requests
+ */
+class UpdateProjectStatusesRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            'title' => 'required',
+        ];
+    }
+}
