+++ app/Http/Requests/UpdateRolesRequest.php
@@ -1,8 +1,11 @@
-<?php
-namespace App\Http\Requests\Admin;
+<?php namespace App\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 
+/**
+ * Class UpdateRolesRequest
+ * @package App\Http\Requests
+ */
 class UpdateRolesRequest extends FormRequest
 {
     /**
