+++ app/Http/Requests/UpdateUsersRequest.php
@@ -1,0 +1,34 @@
+<?php namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+/**
+ * Class UpdateUsersRequest
+ * @package App\Http\Requests
+ */
+class UpdateUsersRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            'name' => 'required',
+            'email' => 'required|email|unique:users,email,'.$this->route('user'),
+            'roles' => 'required',
+        ];
+    }
+}
